Add a way to launch a scene without touching (Platform)Apps #11
All the work is done on Game side by creating and managing scenes with
GameSceneManager (shared/game/gamescenemanager.(h|cpp)).

+++ shared/lib/api/gameobject/sprite.cpp
@@ -47,7 +47,7 @@ void Sprite::draw(IRenderer* renderer) const
 {
     if (m_image)
     {
-        m_image->draw(renderer, m_x, m_y);
+        m_image->draw(renderer, (unsigned int) m_x, (unsigned int) m_y);
     }
 }
